requirements.txt
streamlit
requests
pandas
matplotlib
seaborn
Pillow